+++ experiments/long_runs/snowy_land_pmodel.jl
@@ -93,7 +93,7 @@ function setup_model(
 
     ground = ClimaLand.PrognosticGroundConditions{FT}()
     canopy_forcing = (; atmos, radiation, ground)
-    prognostic_land_components = (:canopy, :snow, :soil, :soilco2)
+    prognostic_land_components = (:canopy, :lake, :snow, :soil, :soilco2)
 
     # Construct the P model manually since it is not a default
     photosynthesis = PModel{FT}(domain, toml_dict)